[Refactor] towns_list をTownList クラスに作り変え、最低限のメソッドを定義した

+++ src/store/store.h
@@ -28,7 +28,7 @@ class PlayerType;
 int16_t store_get_stock_max(StoreSaleType sst, bool powerup = true);
 void store_shuffle(PlayerType *player_ptr, StoreSaleType which);
 void store_maintenance(PlayerType *player_ptr, int town_num, StoreSaleType store_num, int chance);
-void store_init(int town_num, StoreSaleType store_num);
+void store_init(size_t town_num, StoreSaleType store_num);
 void store_examine(PlayerType *player_ptr, StoreSaleType store_num);
 int store_check_num(const ItemEntity *o_ptr, StoreSaleType store_num);
 int store_level(StoreSaleType store_num);
